+++ kernel/tr/updates/insertions.cpp
@@ -68,7 +68,7 @@ void insert_before(PPOpIn arg2, PPOpIn arg1)
                 if (is_node_persistent(node))
 				{
 					tuple tup(2);
-					tup.copy(tuple_cell::node(node),tuple_cell(arg2seq.size()));
+					tup.copy(tuple_cell::node(node),tuple_cell((__int64)(arg2seq.size())));
 					arg3seq.add(tup);
 				}
 				arg2seq.add(indir);	
@@ -308,7 +308,7 @@ void insert_following(PPOpIn arg2, PPOpIn arg1)
 				if (is_node_persistent(node))
 				{
 					tuple tup(2);
-					tup.copy(tuple_cell::node(node),tuple_cell(arg2seq.size()));
+					tup.copy(tuple_cell::node(node),tuple_cell((__int64)(arg2seq.size())));
 					arg3seq.add(tup);
 				}
 				arg2seq.add(indir);					
@@ -537,7 +537,7 @@ void insert_to(PPOpIn arg2, PPOpIn arg1)
 				if (is_node_persistent(node))
 				{
 					tuple tup(2);
-					tup.copy(tuple_cell::node(node),tuple_cell(arg2seq.size()));
+					tup.copy(tuple_cell::node(node),tuple_cell((__int64)(arg2seq.size())));
 					arg3seq.add(tup);
 				}
 				arg2seq.add(indir);
